requirements.txt
anthropic==0.39.0
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy<2,>=1.26.0
streamlit==1.32.0
plotly==5.18.0
streamlit-aggrid==0.3.4.post3